re-styled individual posts to fit styling
