+++ gatsby-node.js
@@ -1,12 +1,12 @@
 const path = require(`path`)
 const { createFilePath } = require(`gatsby-source-filesystem`)
 
-const postsPrefix = 'blog';
+const postsPrefix = "blog"
 
 exports.createPages = ({ graphql, actions }) => {
   const { createPage } = actions
 
-  const blogPost = path.resolve(`./src/templates/blog-post.js`)
+  const blogPost = path.resolve(`./src/pages/blog-post.js`)
   return graphql(
     `
       {
@@ -56,8 +56,6 @@ exports.createPages = ({ graphql, actions }) => {
 
 exports.onCreateNode = ({ node, actions, getNode }) => {
   const { createNodeField } = actions
-
-
 
   if (node.internal.type === `MarkdownRemark`) {
     const value = createFilePath({ node, getNode })
