+++ src/pages/404.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { graphql } from "gatsby"
 
-import Layout from "../components/blog/layout"
+import Layout from "../components/partials/layout"
 import SEO from "../components/seo"
 
 class NotFoundPage extends React.Component {
