+++ src/pages/blog.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { Link, graphql } from "gatsby"
 
-import Layout from "../components/home/layout"
+import Layout from "../components/partials/layout"
 import SEO from "../components/seo"
 import { rhythm } from "../utils/typography"
